Move httpx test case to its own file
