+++ tests/tests38/test_http_aiohttp.py
@@ -1,7 +1,6 @@
 import json
 from unittest import IsolatedAsyncioTestCase
 
-import httpx
 import pytest
 
 from mocket.async_mocket import async_mocketize
@@ -80,22 +79,3 @@ if ENABLE_TEST_CLASS:
                 async with session.get(self.target_url) as get_response:
                     assert get_response.status == 200
                     assert await get_response.text() == '{"origin": "127.0.0.1"}'
-
-
-class HttpxEntryTestCase(IsolatedAsyncioTestCase):
-    target_url = "http://httpbin.local/ip"
-
-    @async_httprettified
-    async def test_httprettish_httpx_session(self):
-        expected_response = {"origin": "127.0.0.1"}
-
-        HTTPretty.register_uri(
-            HTTPretty.GET,
-            self.target_url,
-            body=json.dumps(expected_response),
-        )
-
-        async with httpx.AsyncClient() as client:
-            response = await client.get(self.target_url)
-            assert response.status_code == 200
-            assert response.json() == expected_response
